Stringify cavatica token before saving, move components to cavatica folder
