+++ src/components/Login.js
@@ -63,7 +63,7 @@ export const handleJWT = async ({ jwt, onFinish, setToken, setUser }) => {
 export const fetchIntegrationTokens = ({ setIntegrationToken }) => {
   getCavaticaUser()
     .then(userData => {
-      setIntegrationToken(CAVATICA, userData);
+      setIntegrationToken(CAVATICA, JSON.stringify(userData.data));
     })
     .catch(response => {
       // Could not retrieve cavatica user info, nothing to do.
